Tidies up action driver naming. Moves driver factory into core.
