+++ bot/src/main/kotlin/com/gatehill/corebot/action/DirectActionPerformServiceImpl.kt
@@ -1,17 +1,18 @@
 package com.gatehill.corebot.action
 
+import com.gatehill.corebot.action.driver.ActionDriverFactory
 import com.gatehill.corebot.action.model.PerformActionRequest
 import com.gatehill.corebot.action.model.PerformActionResult
 import java.util.concurrent.CompletableFuture
 import javax.inject.Inject
 
 /**
- * Initiates actions locally by triggering the appropriate driver directly.
+ * Initiates actions by triggering the appropriate driver directly.
  *
  * @author Pete Cornish {@literal <outofcoffee@gmail.com>}
  */
-open class SimpleActionDriverServiceImpl @Inject constructor(
-        private val actionDriverFactory: ActionDriverFactory) : ActionDriverService {
+open class DirectActionPerformServiceImpl @Inject constructor(
+        private val actionDriverFactory: ActionDriverFactory) : ActionPerformService {
 
     override fun perform(request: PerformActionRequest): CompletableFuture<PerformActionResult> {
         val actionDriver = actionDriverFactory.driverFor(request.actionConfig.driver)
