+++ core/src/main/kotlin/com/gatehill/corebot/action/driver/ActionDriver.kt
@@ -1,4 +1,4 @@
-package com.gatehill.corebot.action
+package com.gatehill.corebot.action.driver
 
 import com.gatehill.corebot.action.model.PerformActionResult
 import com.gatehill.corebot.action.model.TriggerContext
