+++ core/src/main/kotlin/com/gatehill/corebot/action/driver/BaseActionDriver.kt
@@ -1,5 +1,7 @@
-package com.gatehill.corebot.action
+package com.gatehill.corebot.action.driver
 
+import com.gatehill.corebot.action.JobTriggerService
+import com.gatehill.corebot.action.LockService
 import com.gatehill.corebot.action.model.PerformActionResult
 import com.gatehill.corebot.action.model.TriggerContext
 import com.gatehill.corebot.chat.model.action.ActionType
